backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
asyncpg
httpx
pydantic[email]
python-dotenv
sentry-sdk
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-jose[cryptography]
python-multipart
resend



# Dev dependencies
pytest
ruff
black
mypy
